+++ packages/cli/src/commands/index.ts
@@ -7,7 +7,7 @@ import { UserConfig } from '../types';
 import { loadUserConfig } from '../utils/loadUserConfig';
 
 export const scripts = (defaultAppConfig: UserConfig) => {
-  const entry = async (): Promise<void> => {
+  const entry = async (): Promise<App> => {
     if (process.env.NODE_ENV === undefined) {
       process.env.NODE_ENV = 'development';
     }
@@ -22,17 +22,17 @@ export const scripts = (defaultAppConfig: UserConfig) => {
       path.resolve(defaultAppConfig.temp, 'config.cjs'),
     ].find((item) => fs.pathExistsSync(item));
 
-    const userConfig = await loadUserConfig(userConfigPath);
+    const { npmConfig = {}, ...userConfig } = await loadUserConfig(userConfigPath);
 
     // resolve the final app config to use
     const appConfig = {
       ...defaultAppConfig,
       ...userConfig,
+      npmConfig: {
+        ...(defaultAppConfig.npmConfig || {}),
+        ...npmConfig,
+      },
     };
-
-    if (appConfig === null) {
-      return;
-    }
 
     // create vuepress app
     const app = new App(appConfig);
@@ -45,6 +45,8 @@ export const scripts = (defaultAppConfig: UserConfig) => {
     // initialize and prepare
     await app.init();
     await app.prepare();
+
+    return app;
   };
 
   return entry;